+++ resources/styles.css
@@ -769,10 +769,6 @@ select {
   pointer-events: auto;
 }
 
-.\!visible {
-  visibility: visible !important;
-}
-
 .visible {
   visibility: visible;
 }
@@ -869,12 +865,30 @@ select {
   z-index: 50;
 }
 
-.ml-\[14px\] {
-  margin-left: 14px;
+.my-2 {
+  margin-top: 0.5rem;
+  margin-bottom: 0.5rem;
+}
+
+.my-\[8px\] {
+  margin-top: 8px;
+  margin-bottom: 8px;
+}
+
+.ml-5 {
+  margin-left: 1.25rem;
+}
+
+.mr-\[8px\] {
+  margin-right: 8px;
 }
 
 .mt-2 {
   margin-top: 0.5rem;
+}
+
+.box-border {
+  box-sizing: border-box;
 }
 
 .block {
@@ -891,6 +905,10 @@ select {
 
 .hidden {
   display: none;
+}
+
+.h-2 {
+  height: 0.5rem;
 }
 
 .h-3 {
@@ -930,8 +948,8 @@ select {
   height: 100%;
 }
 
-.max-h-\[750px\] {
-  max-height: 750px;
+.min-h-\[30px\] {
+  min-height: 30px;
 }
 
 .min-h-\[40px\] {
@@ -940,6 +958,10 @@ select {
 
 .min-h-\[80px\] {
   min-height: 80px;
+}
+
+.w-2 {
+  width: 0.5rem;
 }
 
 .w-3 {
@@ -973,10 +995,6 @@ select {
 
 .min-w-\[150px\] {
   min-width: 150px;
-}
-
-.max-w-sm {
-  max-width: 24rem;
 }
 
 .grow {
@@ -1017,8 +1035,16 @@ select {
   flex-direction: column;
 }
 
+.items-start {
+  align-items: flex-start;
+}
+
 .items-center {
   align-items: center;
+}
+
+.justify-start {
+  justify-content: flex-start;
 }
 
 .justify-end {
@@ -1027,6 +1053,15 @@ select {
 
 .justify-center {
   justify-content: center;
+}
+
+.justify-between {
+  justify-content: space-between;
+}
+
+.gap-x-1 {
+  -moz-column-gap: 0.25rem;
+       column-gap: 0.25rem;
 }
 
 .gap-x-2 {
@@ -1039,6 +1074,11 @@ select {
        column-gap: 0.75rem;
 }
 
+.gap-x-4 {
+  -moz-column-gap: 1rem;
+       column-gap: 1rem;
+}
+
 .gap-y-2 {
   row-gap: 0.5rem;
 }
@@ -1049,6 +1089,10 @@ select {
 
 .overflow-auto {
   overflow: auto;
+}
+
+.whitespace-nowrap {
+  white-space: nowrap;
 }
 
 .rounded {
@@ -1093,6 +1137,11 @@ select {
   border-color: rgb(239 68 68 / var(--tw-border-opacity));
 }
 
+.border-red-600 {
+  --tw-border-opacity: 1;
+  border-color: rgb(220 38 38 / var(--tw-border-opacity));
+}
+
 .border-transparent {
   border-color: transparent;
 }
@@ -1117,12 +1166,25 @@ select {
   background-color: var(--secondary-color-120);
 }
 
+.bg-red-600 {
+  --tw-bg-opacity: 1;
+  background-color: rgb(220 38 38 / var(--tw-bg-opacity));
+}
+
 .bg-transparent {
   background-color: transparent;
 }
 
+.p-1 {
+  padding: 0.25rem;
+}
+
 .p-2 {
   padding: 0.5rem;
+}
+
+.p-3 {
+  padding: 0.75rem;
 }
 
 .p-4 {
@@ -1154,12 +1216,21 @@ select {
   padding-bottom: 0.75rem;
 }
 
+.py-\[5px\] {
+  padding-top: 5px;
+  padding-bottom: 5px;
+}
+
 .pb-4 {
   padding-bottom: 1rem;
 }
 
 .pl-4 {
   padding-left: 1rem;
+}
+
+.pl-\[22px\] {
+  padding-left: 22px;
 }
 
 .pr-4 {
@@ -1176,6 +1247,10 @@ select {
 
 .text-2xl {
   font-size: 1.5rem;
+}
+
+.text-base {
+  font-size: 1.1rem;
 }
 
 .text-lg {
@@ -1204,11 +1279,6 @@ select {
 
 .leading-6 {
   line-height: 1.5rem;
-}
-
-.text-gray-300 {
-  --tw-text-opacity: 1;
-  color: rgb(209 213 219 / var(--tw-text-opacity));
 }
 
 .text-gray-400 {
@@ -1341,27 +1411,6 @@ select {
   border-radius: 5px;
 }
 
-.tooeen-tree-item {
-  display: flex;
-  flex-direction: column;
-}
-
-.tooeen-tree-item-title {
-  display: flex;
-  align-items: center;
-  min-height: 30px;
-  width: 100%;
-  box-sizing: border-box;
-  padding-right: 10px;
-  cursor: pointer;
-}
-
-.tooeen-tree-item-title:hover {
-  background-color: var(--primary-color);
-  color: var(--secondary-color-100);
-  border-radius: 5px;
-}
-
 .placeholder\:text-gray-400::-moz-placeholder {
   --tw-text-opacity: 1;
   color: rgb(156 163 175 / var(--tw-text-opacity));
@@ -1380,9 +1429,9 @@ select {
   border-color: var(--primary-color);
 }
 
-.hover\:bg-green-500:hover {
+.hover\:bg-\[\#55A014\]:hover {
   --tw-bg-opacity: 1;
-  background-color: rgb(34 197 94 / var(--tw-bg-opacity));
+  background-color: rgb(85 160 20 / var(--tw-bg-opacity));
 }
 
 .hover\:bg-ifcjs-120:hover {
