+++ src/core/index.ts
@@ -8,9 +8,7 @@ export * from "./SimpleGrid";
 export * from "./Components";
 export * from "./SimpleClipper";
 export * from "./ScreenCuller";
-export * from "./SimpleDimensions";
 export * from "./LocalCacher";
-export * from "./SimpleAngle";
 export * from "./SimpleSVGViewport";
 export * from "./Simple2DMarker";
 export * from "./MaterialManager";
