+++ src/ifc/IfcPropertiesProcessor/src/index.ts
@@ -1,5 +1,7 @@
 export * from "./element-psets";
 export * from "./element-qsets";
-export * from "./pset-props";
-export * from "./qset-quantities";
-export * from "./relation-map";
+export * from "./element-storey";
+export * from "./edit-prop";
+export * from "./new-prop";
+export * from "./new-pset";
+export * from "./property-tag";
