+++ src/measurement/index.ts
@@ -2,3 +2,4 @@ export * from "./AreaMeasureElement";
 export * from "./AreaMeasurement";
 export * from "./AngleMeasureElement";
 export * from "./AngleMeasurement";
+export * from "./LengthMeasurement";
